Update generator.js
The generator now supports multiple view directories instead of just one. This is much like handlebars in express normally works. It will also throw and error that is using same params as a normal file not found if no file is found for either template or text template which is how it would act before this change. Very transparent.